+++ src/org/minima/kissvm/functions/MinimaFunction.java
@@ -13,15 +13,15 @@ import org.minima.kissvm.functions.cast.BOOL;
 import org.minima.kissvm.functions.cast.HEX;
 import org.minima.kissvm.functions.cast.NUMBER;
 import org.minima.kissvm.functions.cast.STRING;
+import org.minima.kissvm.functions.hex.BITCOUNT;
+import org.minima.kissvm.functions.hex.BITGET;
+import org.minima.kissvm.functions.hex.BITSET;
 import org.minima.kissvm.functions.hex.CONCAT;
 import org.minima.kissvm.functions.hex.GET;
 import org.minima.kissvm.functions.hex.LEN;
 import org.minima.kissvm.functions.hex.REV;
 import org.minima.kissvm.functions.hex.SUBSET;
 import org.minima.kissvm.functions.number.ABS;
-import org.minima.kissvm.functions.number.BITCOUNT;
-import org.minima.kissvm.functions.number.BITGET;
-import org.minima.kissvm.functions.number.BITSET;
 import org.minima.kissvm.functions.number.CEIL;
 import org.minima.kissvm.functions.number.DEC;
 import org.minima.kissvm.functions.number.FLOOR;
