+++ src/org/minima/kissvm/functions/hex/BITGET.java
@@ -1,20 +1,20 @@
-package org.minima.kissvm.functions.number;
+package org.minima.kissvm.functions.hex;
 
 import java.util.BitSet;
 
 import org.minima.kissvm.Contract;
 import org.minima.kissvm.exceptions.ExecutionException;
 import org.minima.kissvm.functions.MinimaFunction;
-import org.minima.kissvm.values.HexValue;
+import org.minima.kissvm.values.BooleanValue;
 import org.minima.kissvm.values.Value;
 
-public class BITSET extends MinimaFunction {
+public class BITGET extends MinimaFunction {
 
 	/**
 	 * @param zName
 	 */
-	public BITSET() {
-		super("BITSET");
+	public BITGET() {
+		super("BITGET");
 	}
 	
 	/* (non-Javadoc)
@@ -22,7 +22,7 @@ public class BITSET extends MinimaFunction {
 	 */
 	@Override
 	public Value runFunction(Contract zContract) throws ExecutionException {
-		checkExactParamNumber(3);
+		checkExactParamNumber(2);
 		
 		//get the Input Data
 		byte[] data = zContract.getHexParam(0, this).getRawData();
@@ -31,27 +31,21 @@ public class BITSET extends MinimaFunction {
 		//Get the desired Bit
 		int bit = zContract.getNumberParam(1, this).getNumber().getAsInt();
 		if(bit<0 || bit>totbits) {
-			throw new ExecutionException("BitSet too large "+bit+" / "+totbits);
+			throw new ExecutionException("BitGet too large "+bit+" / "+totbits);
 		}
 		
-		//Set to ON or OFF
-		boolean set = zContract.getBoolParam(2, this).isTrue();
-
 		//Create a BitSet object..
 		BitSet bits = BitSet.valueOf(data);
 		
-		//Change the bit
-		bits.set(bit, set);
+		//Now check the value..
+		boolean isSet = bits.get(bit);
 		
-		//Now get the new byte array
-		byte[] newarray = bits.toByteArray();
-		
-		//Return the new HEXValue
-		return new HexValue(newarray);		
+		//return..
+		return new BooleanValue(isSet);
 	}
 	
 	@Override
 	public MinimaFunction getNewFunction() {
-		return new BITSET();
+		return new BITGET();
 	}
 }
