+++ test/org/minima/tests/kissvm/functions/maths/BITCOUNTTests.java
@@ -12,7 +12,7 @@ import org.minima.kissvm.exceptions.ExecutionException;
 import org.minima.kissvm.exceptions.MinimaParseException;
 import org.minima.kissvm.expressions.ConstantExpression;
 import org.minima.kissvm.functions.MinimaFunction;
-import org.minima.kissvm.functions.number.BITCOUNT;
+import org.minima.kissvm.functions.hex.BITCOUNT;
 import org.minima.kissvm.values.BooleanValue;
 import org.minima.kissvm.values.HexValue;
 import org.minima.kissvm.values.NumberValue;
